[#5] Minor touch-up.  Changed the return types for getMaximumIntensity from long to double.  This will do better to match the result from getFormattedSpectrum, and it makes it possible for future spectrometers to provide pre-processed spectra with non-integer intensity values.

+++ branches/amg-merge-5/SeaBreeze/include/api/seabreezeapi/DeviceAdapter.h
@@ -101,7 +101,7 @@ namespace seabreeze {
                     unsigned long integrationTimeMicros);
             unsigned long spectrometerGetMinimumIntegrationTimeMicros(
                     long spectrometerFeatureID, int *errorCode);
-            unsigned long spectrometerGetMaximumIntensity(
+            double spectrometerGetMaximumIntensity(
                     long spectrometerFeatureID, int *errorCode);
             int spectrometerGetUnformattedSpectrumLength(
                     long spectrometerFeatureID, int *errorCode);
