+++ branches/amg-merge-5/SeaBreeze/include/api/seabreezeapi/SeaBreezeAPI.h
@@ -133,7 +133,7 @@ public:
     void spectrometerSetTriggerMode(long deviceID, long spectrometerFeatureID, int *errorCode, int mode);
     void spectrometerSetIntegrationTimeMicros(long deviceID, long spectrometerFeatureID, int *errorCode, unsigned long integrationTimeMicros);
     unsigned long spectrometerGetMinimumIntegrationTimeMicros(long deviceID, long spectrometerFeatureID, int *errorCode);
-    unsigned long spectrometerGetMaximumIntensity(long deviceID, long spectrometerFeatureID, int *errorCode);
+    double spectrometerGetMaximumIntensity(long deviceID, long spectrometerFeatureID, int *errorCode);
     int spectrometerGetUnformattedSpectrumLength(long deviceID, long spectrometerFeatureID, int *errorCode);
     int spectrometerGetUnformattedSpectrum(long deviceID, long spectrometerFeatureID, int *errorCode, unsigned char *buffer, int bufferLength);
     int spectrometerGetFormattedSpectrumLength(long deviceID, long spectrometerFeatureID, int *errorCode);
@@ -748,7 +748,7 @@ extern "C" {
             long featureID, int *error_code);
 
     /**
-     * This function returns the maximum pixel intensity, for the
+     * This function returns the maximum pixel intensity for the
      * spectrometer.
      *
      * @param deviceID (Input) The index of a device previously opened with
@@ -758,10 +758,10 @@ extern "C" {
      *      sbapi_get_spectrometer_features() function.
      * @param error_code (Output) A pointer to an integer that can be used
      *      for storing error codes.
-     * @return Returns maximum pixel intensity time if > 0.
+     * @return Returns maximum pixel intensity if > 0.
      *      On error, returns -1 and error_code will be set accordingly.
      */
-    DLL_DECL long
+    DLL_DECL double
     sbapi_spectrometer_get_maximum_intensity(long deviceID,
             long featureID, int *error_code);
