+++ branches/amg-merge-5/SeaBreeze/include/api/seabreezeapi/SpectrometerFeatureAdapter.h
@@ -64,7 +64,7 @@ namespace seabreeze {
             void setIntegrationTimeMicros(int *errorCode,
                     unsigned long integrationTimeMicros);
             long getMinimumIntegrationTimeMicros(int *errorCode);
-            long getMaximumIntensity(int *errorCode);
+            double getMaximumIntensity(int *errorCode);
         };
 
     }
